find_word/requirements.txt
ffmpeg-python==0.2.0
future==0.18.2
numpy==1.17.4
pandas==0.25.3
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
srt==3.0.0
